+++ frontend/src/app/main/data/workspace/bool.cljs
@@ -6,10 +6,9 @@
 
 (ns app.main.data.workspace.bool
   (:require
-   [app.common.colors :as clr]
    [app.common.data :as d]
    [app.common.geom.shapes :as gsh]
-   [app.common.pages.changes-builder :as cb]
+   [app.common.pages.changes-builder :as pcb]
    [app.common.pages.helpers :as cph]
    [app.common.path.shapes-to-path :as stp]
    [app.common.uuid :as uuid]
@@ -35,8 +34,8 @@
   (let [shapes (mapv #(stp/convert-to-path % objects) shapes)
         head (if (= bool-type :difference) (first shapes) (last shapes))
         head (cond-> head
-               (and (contains? head :svg-attrs) (nil? (:fill-color head)))
-               (assoc :fill-color clr/black))
+               (and (contains? head :svg-attrs) (empty? (:fills head)))
+               (assoc :fills stp/default-bool-fills))
 
         head-data (select-keys head stp/style-properties)
 
@@ -61,8 +60,8 @@
                     (mapv #(stp/convert-to-path % objects)))
         head (if (= bool-type :difference) (first shapes) (last shapes))
         head (cond-> head
-               (and (contains? head :svg-attrs) (nil? (:fill-color head)))
-               (assoc :fill-color clr/black))
+               (and (contains? head :svg-attrs) (empty? (:fills head)))
+               (assoc :fills stp/default-bool-fills))
         head-data (select-keys head stp/style-properties)]
 
     (-> group
@@ -97,10 +96,10 @@
         (when-not (empty? shapes)
           (let [[boolean-data index] (create-bool-data bool-type name shapes objects)
                 shape-id (:id boolean-data)
-                changes (-> (cb/empty-changes it page-id)
-                            (cb/with-objects objects)
-                            (cb/add-obj boolean-data {:index index})
-                            (cb/change-parent shape-id shapes))]
+                changes (-> (pcb/empty-changes it page-id)
+                            (pcb/with-objects objects)
+                            (pcb/add-obj boolean-data {:index index})
+                            (pcb/change-parent shape-id shapes))]
             (rx/of (dch/commit-changes changes)
                    (dwc/select-shapes (d/ordered-set shape-id)))))))))