+++ frontend/src/app/main/render.cljs
@@ -215,9 +215,34 @@
              [:& shape-wrapper {:shape item
                                 :key (:id item)}])))]]]))
 
+(mf/defc file-thumbnail-svg
+  {::mf/wrap [mf/memo]}
+  [{:keys [data embed? include-metadata?] :as props
+    :or {embed? false include-metadata? false}}]
+  (let [data (assoc data :x 0 :y 0)
+        vbox   (format-viewbox {:width (:width data 0) :height (:height data 0)})
+        background-color (get-in data [:options :background] default-color)]
+
+    [:& (mf/provider embed/context) {:value embed?}
+     [:& (mf/provider export/include-metadata-ctx) {:value include-metadata?}
+      [:svg {:view-box vbox
+             :version "1.1"
+             :xmlns "http://www.w3.org/2000/svg"
+             :xmlnsXlink "http://www.w3.org/1999/xlink"
+             :xmlns:penpot (when include-metadata? "https://penpot.app/xmlns")
+             :style {:width "100%"
+                     :height "100%"
+                     :background background-color}}
+
+       (when include-metadata?
+         [:& export/export-page {:options (:options data)}])
+
+       [:> shape-container {:shape data}
+        [:& frame/frame-thumbnail {:shape data}]]]]]))
+
 (mf/defc frame-svg
   {::mf/wrap [mf/memo]}
-  [{:keys [objects frame zoom] :or {zoom 1} :as props}]
+  [{:keys [objects frame zoom show-thumbnails?] :or {zoom 1} :as props}]
   (let [frame-id          (:id frame)
         include-metadata? (mf/use-ctx export/include-metadata-ctx)
 
@@ -253,7 +278,13 @@
            :xmlns "http://www.w3.org/2000/svg"
            :xmlnsXlink "http://www.w3.org/1999/xlink"
            :xmlns:penpot (when include-metadata? "https://penpot.app/xmlns")}
-     [:& wrapper {:shape frame :view-box vbox}]]))
+     (if (or (not show-thumbnails?) (nil? (:thumbnail frame)))
+       [:& wrapper {:shape frame :view-box vbox}]
+
+       ;; Render the frame thumbnail
+       (let [frame (gsh/transform-shape frame)]
+         [:> shape-container {:shape frame}
+          [:& frame/frame-thumbnail {:shape frame}]]))]))
 
 (mf/defc component-svg
   {::mf/wrap [mf/memo #(mf/deferred % ts/idle-then-raf)]}