+++ frontend/src/app/main/ui/workspace/viewport.cljs
@@ -266,12 +266,11 @@
            :zoom zoom}])
 
        (when show-selection-handlers?
-         [:& selection/selection-handlers
-          {:selected selected
-           :shapes selected-shapes
+         [:& selection/selection-area
+          {:shapes selected-shapes
            :zoom zoom
            :edition edition
-           :disable-handlers (or drawing-tool edition @space?)
+           :disable-handlers (or drawing-tool edition @space? @ctrl?)
            :on-move-selected on-move-selected
            :on-context-menu on-menu-selected}])
 
@@ -378,16 +377,33 @@
          :viewport-ref viewport-ref}]
 
        (when show-rules?
-         [:*
-          [:& rules/rules
-           {:zoom zoom
-            :vbox vbox
-            :selected-shapes selected-shapes}]
+         [:& rules/rules
+          {:zoom zoom
+           :vbox vbox
+           :selected-shapes selected-shapes}])
 
-          [:& guides/viewport-guides
-           {:zoom zoom
-            :vbox vbox
-            :hover-frame frame-parent
-            :modifiers modifiers
-            :disabled-guides? disabled-guides?}]])]]]))
+       (when show-rules?
+         [:& guides/viewport-guides
+          {:zoom zoom
+           :vbox vbox
+           :hover-frame frame-parent
+           :modifiers modifiers
+           :disabled-guides? disabled-guides?}])
 
+       (when show-selection-handlers?
+         [:g.selection-handlers {:clipPath "url(#clip-handlers)"}
+          [:defs
+           (let [rule-area-size (/ rules/rule-area-size zoom)]
+             ;; This clip is so the handlers are not over the rules
+             [:clipPath {:id "clip-handlers"}
+              [:rect {:x (+ (:x vbox) rule-area-size)
+                      :y (+ (:y vbox) rule-area-size)
+                      :width (- (:width vbox) (* rule-area-size 2))
+                      :height (- (:height vbox) (* rule-area-size 2))}]])]
+
+          [:& selection/selection-handlers
+           {:selected selected
+            :shapes selected-shapes
+            :zoom zoom
+            :edition edition
+            :disable-handlers (or drawing-tool edition @space?)}]])]]]))